+++ src/global.d.ts
@@ -1,0 +1,4 @@
+declare module '@fastly/cli' {
+  const cli: string;
+  export default cli;
+}