+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- feat: Use fastly command from `@fastly/cli` package
+
 ### Fixed
 
 - stream error output from fastly command